svelte-migrate routes script
