+++ src/routes/rss.xml/+server.js
@@ -1,3 +1,4 @@
+import { json } from '@sveltejs/kit';
 import { description, siteBaseUrl, title } from "$lib/meta";
 import posts from '$lib/posts';
 
@@ -7,6 +8,11 @@ export async function GET() {
 		'Cache-Control': 'max-age=0, s-maxage=3600',
 		'Content-Type': 'application/xml'
 	};
+	throw new Error("@migration task: Migrate this return statement (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292701)");
+	// Suggestion (check for correctness before using):
+	// return json(body, {
+	//   headers: headers
+	// });
 	return {
 		headers,
 		body
